requirements/base.txt
Django==2.2.5
djangorestframework==3.10.2
django-rest-swagger==2.2.0
drf-yasg==1.16.1
django-cors-headers-3.1.0
psycopg2==2.8.3
python-socketio-4.3.1
eventlet==0.25.1
gevent-1.4.0